# lxdapi/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "魔方财务系统专用LXD容器管理后端API服务",
        "title": "魔方财务-LXD对接API",
        "termsOfService": "https://github.com/xkatld/zjmf-lxd-server",
        "contact": {
            "name": "API支持",
            "url": "https://github.com/xkatld/zjmf-lxd-server",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/xkatld/zjmf-lxd-server/blob/main/LICENSE"
        },
        "version": "1.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "通过魔方财务创建 LXD 容器",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "容器管理"
                ],
                "summary": "创建容器",
                "parameters": [
                    {
                        "description": "创建参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LXDServerCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.LXDServerResponse"
                        }
                    },
                    "401": {
                        "description": "认证失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/password": {
            "post": {
                "description": "提交密码重置任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "容器管理"
                ],
                "summary": "重置密码",
                "parameters": [
                    {
                        "description": "密码参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LXDServerResetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.LXDServerResponse"
                        }
                    }
                }
            }
        },
...
        "/api/reinstall": {
            "post": {
                "description": "提交容器重装任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "容器管理"
                ],
                "summary": "重装容器系统",
                "parameters": [
                    {
                        "description": "重装参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LXDServerReinstallRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.LXDServerResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.LXDServerCreateRequest": {
            "type": "object",
            "required": [
                "hostname"
            ],
            "properties": {
                "allow_nesting": {
                    "description": "允许嵌套",
                    "type": "boolean",
                    "example": false
                },
                "cpu_allowance": {
                    "description": "CPU 高级配置",
                    "type": "string",
                    "example": "50%"
                },
                "cpus": {
                    "type": "integer",
                    "example": 2
                },
                "disk": {
                    "description": "硬盘(支持单位)",
                    "type": "string",
                    "example": "512MB"
                },
                "disk_io_limit": {
                    "description": "磁盘IO限制",
                    "type": "string",
                    "example": "100MB"
                },
                "egress": {
                    "description": "下行带宽(支持单位) - 容器接收数据限制",
                    "type": "string",
                    "example": "100Mbit"
                },
                "enable_lxcfs": {
                    "description": "LXCFS 配置",
                    "type": "boolean",
                    "example": true
                },
                "hostname": {
                    "description": "容器主机名（对应容器名）",
                    "type": "string",
                    "example": "test-container"
                },
                "image": {
                    "description": "镜像",
                    "type": "string",
                    "example": "debian/12"
                },
                "ingress": {
                    "description": "上行带宽(支持单位) - 容器发送数据限制",
                    "type": "string",
                    "example": "100Mbit"
                },
                "max_processes": {
                    "description": "进程和文件限制",
                    "type": "integer",
                    "example": 512
                },
                "memory": {
                    "description": "内存(支持单位)",
                    "type": "string",
                    "example": "256MB"
                },
                "memory_swap": {
                    "description": "内存高级配置",
                    "type": "boolean",
                    "example": true
                },
                "network_mode": {
                    "type": "string",
                    "example": "mode1"
                },
                "password": {
                    "description": "root密码",
                    "type": "string",
                    "example": "password123"
                },
                "privileged": {
                    "description": "是否允许特权模式",
                    "type": "boolean",
                    "example": false
                },
                "traffic_limit": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "models.LXDServerReinstallRequest": {
            "type": "object",
            "required": [
                "hostname",
                "system"
            ],
            "properties": {
                "allow_nesting": {
                    "type": "boolean",
                    "example": false
                },
                "cpu_allowance": {
                    "description": "CPU 高级配置",
                    "type": "string",
                    "example": "50%"
                },
                "cpus": {
                    "description": "基础配置（重装时重新应用）",
                    "type": "integer",
                    "example": 2
                },
                "disk": {
                    "type": "string",
                    "example": "512MB"
                },
                "disk_io_limit": {
                    "description": "磁盘IO限制",
                    "type": "string",
                    "example": "100MB"
                },
                "egress": {
                    "type": "string",
                    "example": "100Mbit"
                },
                "enable_lxcfs": {
                    "description": "LXCFS 配置",
                    "type": "boolean",
                    "example": true
                },
                "hostname": {
                    "description": "容器主机名",
                    "type": "string",
                    "example": "test-container"
                },
                "ingress": {
                    "type": "string",
                    "example": "100Mbit"
                },
                "max_processes": {
                    "description": "进程和文件限制",
                    "type": "integer",
                    "example": 512
                },
                "memory": {
                    "type": "string",
                    "example": "256MB"
                },
                "memory_swap": {
                    "description": "内存高级配置",
                    "type": "boolean",
                    "example": true
                },
                "password": {
                    "description": "新的root密码",
                    "type": "string",
                    "example": "newpassword123"
                },
                "privileged": {
                    "type": "boolean",
                    "example": false
                },
                "system": {
                    "description": "新的操作系统镜像",
                    "type": "string",
                    "example": "debian/12"
                },
                "traffic_limit": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "models.LXDServerResetPasswordRequest": {
            "type": "object",
            "required": [
                "hostname",
                "password"
            ],
            "properties": {
                "hostname": {
                    "type": "string",
                    "example": "test-container"
                },
                "password": {
                    "type": "string",
                    "example": "newpassword123"
                }
            }
        },
        "models.LXDServerResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "在header中添加apikey字段，值为配置文件中的api_hash",
            "type": "apiKey",
            "name": "apikey",
            "in": "header"
        }
    }
}